fakultaet.c erstellt

+++ fakultaet.c
@@ -6,7 +6,7 @@ int main(int argc, char ** argv) {
 int b; 
 int c=1; 
 int a;
-        a = atof(argv[1]);
+        a = atoi(argv[1]);
 if(a>0){
 
 b=a;
